First of many
Actual description of what's happening, just this once. I'm separating the project into the engine core, which will become a static library, the editor, and the runtime, which both will be executables. The editor is self-explanatory, it will function as the interface the user uses to actually edit the scenes in the game. The runtime will be a minimal "bootstrap" type of application, which all the user-written components can hook onto. How that will be achieved is currently beyond me but whatever, I got the linking to work so that's a huge W